Convert shader effect to primitive "poplights" without persistence.
Shader now emits particles for each mesh triangle.  These particles are emitted probabilistically based on the time elapsed since the last frame was drawn, the surface area of the triangle, and the specified frequency of emission per square 'unit'.

When emitted, a particle is positioned randomly on the face of its' triangle's face, and assigned a random altitude between 'minAltitude' and 'maxAltitude'.  The particle is also assigned a random radius between 'minRadius' and 'maxRadius'.  This radius will define the maximum radius for the particle during its lifetime.

NOTE: This shader currently uses an absolutely terrible noise generator.  It requires the use of a seed uniform (provided by the CPU) and arbitrary input values (such as hardcoded constants, vertex position, polygon normal, etc.) to produce a single value in [0,1).